Fix invalid test tag
Remove the '.' from the tag name as dots are invalid in resource names.
```
The ProjectRequest "runtime-operator-test-v0.8.0-rc2-503" is invalid:
metadata.name: Invalid value: "runtime-operator-test-v0.8.0-rc2-503": a
DNS-1123 label must consist of lower case alphanumeric characters or
'-', and must start and end with an alphanumeric character (e.g.
'my-name',  or '123-abc', regex used for validation is
'[a-z0-9]([-a-z0-9]*[a-z0-9])?')
```