Dont create deployment without a pull secret
The cluster automatically creates pull secrets for the service account.
This can take a while.
Sometimes, the controller creates the deployment before the pull secret
is created. This causes the pods to go into an image pull backoff loop.
Check that the pull secret exists before the deployment is created, to
prevent this