Ensure pods are updated if the service account is (#387)
* Delete stale pull secrets from service account

Previously, if a pull secret was removed from the CR, it could not
be removed from the service account, as it was impossible to know which
secret to remove.
Now, the secret is tracked in the status, so it can be removed from
the service account if it is removed from the CR.

* Ensure pods are updated if the service account is

This fixes an issue where, after adding a pull secret to a
runtime component, the pods were not updated with the secret.
The pods had to be manually deleted in order to pick up the change
Fix some tests to deal with this, as the pods now may get updated
more often

* Review comments

Set the service account reference for all service accounts
and unset the reference if the pull secret is removed

* Use a better way to add pull secrets to the SA

Currently, the pull secret is always added at index 0 to ensure it
is only added once. This is fragile. Instead, check if the pull
secret already exists, and if not, then append it. This still
ensures that the secret will only be added once

* Fix tests

A pull secret will now be appended rather than set at index 0, so
fix up the tests to match this behaviour.

* Refactor probe tests to not use observedGeneration

* Refactor affinity tests to not use observedGen

Co-authored-by: Hank W Ibell <hwibell@us.ibm.com>