Use reader client to get operator config map (#591)
The operator config map is not always in the same namespace that is
being watched, and the client provided by the manager will only read
resources in the watched namespace. 'Get' on the configmap will fail
in this situation

Use the reader client instead, as this will read resources from any
namespace

Also fix up tests. Need to use the same fake client to create and fetch the configmap
otherwise, the configmap can't be retrieved